Continue refactoring the bluetooth functions
